# perf: Use `ctz` instruction and `Lsb0`