This PR improves perfomance of `calculate_gaps` and `find_available_blocks` even further by leveraging two things:

- faster creation of bitvec from bitslice. And explicitly choosing `Lsb0` (At least for my system... We could check with a different bitslice and different bit order)
- replacing loops of shifting a single bit at a time with using `count trailing zeros` instruction and shifting a larger amount of bits at once.


Now the "bottleneck" (it takes a little more than half of the operation time) would be the bitvec creation, but I don't see how we could avoid that and still be able to process the slice as its raw representation.

## Results
Before:
```
------------ 1 thread(s) -------------
**read_heavy** with prefill_fraction 0.95
PayloadStorage:
1572864 operations across 1 thread(s) in 1.601686375s; time/op = 1.018µs
RocksDB:
1572864 operations across 1 thread(s) in 905.831959ms; time/op = 575ns

**insert_heavy** with prefill_fraction 0.0
PayloadStorage:
1572864 operations across 1 thread(s) in 10.361096084s; time/op = 6.587µs
RocksDB:
1572864 operations across 1 thread(s) in 3.359631417s; time/op = 2.135µs

**update_heavy** with prefill_fraction 0.5
PayloadStorage:
1572864 operations across 1 thread(s) in 8.492014708s; time/op = 5.399µs
RocksDB:
1572864 operations across 1 thread(s) in 1.965619542s; time/op = 1.249µs
```

After:
```
------------ 1 thread(s) -------------
**read_heavy** with prefill_fraction 0.95
PayloadStorage:
1572864 operations across 1 thread(s) in 946.685667ms; time/op = 601ns
RocksDB:
1572864 operations across 1 thread(s) in 954.203375ms; time/op = 606ns

**insert_heavy** with prefill_fraction 0.0
PayloadStorage:
1572864 operations across 1 thread(s) in 2.717830917s; time/op = 1.727µs
RocksDB:
1572864 operations across 1 thread(s) in 3.521887125s; time/op = 2.239µs

**update_heavy** with prefill_fraction 0.5
PayloadStorage:
1572864 operations across 1 thread(s) in 1.952809875s; time/op = 1.241µs
RocksDB:
1572864 operations across 1 thread(s) in 1.97704325s; time/op = 1.256µs
```